Moves BlobHttpContent into http package.
