+++ crepecake/src/main/java/com/google/cloud/tools/crepecake/http/Request.java
@@ -21,7 +21,6 @@ import com.google.api.client.http.HttpHeaders;
 import com.google.api.client.http.HttpRequestFactory;
 import com.google.api.client.http.javanet.NetHttpTransport;
 import com.google.cloud.tools.crepecake.blob.Blob;
-import com.google.cloud.tools.crepecake.blob.Blobs;
 import com.google.common.annotations.VisibleForTesting;
 import java.io.IOException;
 import java.net.URL;
@@ -72,11 +71,11 @@ public class Request {
 
   /** Sends the request with method POST. */
   public Response post(Blob body) throws IOException {
-    return new Response(requestFactory.buildPostRequest(url, Blobs.toHttpContent(body)));
+    return new Response(requestFactory.buildPostRequest(url, new BlobHttpContent(body)));
   }
 
   /** Sends the request with method PUT. */
   public Response put(Blob body) throws IOException {
-    return new Response(requestFactory.buildPutRequest(url, Blobs.toHttpContent(body)));
+    return new Response(requestFactory.buildPutRequest(url, new BlobHttpContent(body)));
   }
 }
